pyproject.toml
 [project]
name = "agent-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antlr4-python3-runtime==4.9.3",
    "asyncmy>=0.2.11",
    "cryptography>=46.0.5",
    "elasticsearch[async]>=8,<9",
    "fastapi[standard]>=0.135.1",
    "jieba>=0.42.1",
    "langchain>=1.2.12",
    "langchain-huggingface>=1.2.1",
    "langgraph>=1.1.3",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "sqlalchemy>=2.0.48",
    "langchain-openai"
]
